python/insight/external/influxdb-mcp/pyproject.toml
[project]
name = "influxdb-v1-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.2",
    "requests",
    "python-dotenv>=1.0.1",
]
